Don't hard code the project ID.
The project ID is available in the credential file.  If a credential
file is supplied, the project ID will be taken from there.

According to the documentation for FirestoreDb.Create, if the project
ID is passed as null, the library will attempt to auto-detect it.
Hopefully this means it will query the metadata server, if the program
is running on Google infrastructure.